+++ ControleDespesas_ApiNetCore21_Dapper/ControleDespesas.Domain/Empresas/Interfaces/Handlers/IEmpresaHandler.cs
@@ -7,7 +7,7 @@ namespace ControleDespesas.Domain.Empresas.Interfaces.Handlers
     public interface IEmpresaHandler
     {
         ICommandResult<Notificacao> Handler(AdicionarEmpresaCommand command);
-        ICommandResult<Notificacao> Handler(int id, AtualizarEmpresaCommand command);
-        ICommandResult<Notificacao> Handler(int id);
+        ICommandResult<Notificacao> Handler(long id, AtualizarEmpresaCommand command);
+        ICommandResult<Notificacao> Handler(long id);
     }
 }
