+++ ControleDespesas_ApiNetCore21_Dapper/ControleDespesas.Test/AppConfigurations/QueriesSQL/CREATE.sql
@@ -64,7 +64,7 @@ END
 IF NOT EXISTS (SELECT 1 FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_TYPE='BASE TABLE' AND TABLE_NAME='Empresa') 
 BEGIN
     CREATE TABLE [dbo].[Empresa] (
-        [Id] [int] IDENTITY(1,1) NOT NULL,
+        [Id] [bigint] IDENTITY(1,1) NOT NULL,
         [Nome] [nvarchar](100) NOT NULL,
         [Logo] [text] NOT NULL,
         CONSTRAINT [PK_Empresa] PRIMARY KEY CLUSTERED 
@@ -119,7 +119,7 @@ BEGIN
 	CREATE TABLE [dbo].[Pagamento](
 		[Id] [int] IDENTITY(1,1) NOT NULL,
 		[IdTipoPagamento] [int] NOT NULL,
-		[IdEmpresa] [int] NOT NULL,
+		[IdEmpresa] [bigint] NOT NULL,
 		[IdPessoa] [int] NOT NULL,
 		[Descricao] [nvarchar](250) NOT NULL,
 		[Valor] [money] NOT NULL,
